==> statistical_ner_labelling/notebooks/pipeline/1_entity_separation.ipynb <==
## Entity separation

This notebook takes the CSV file from the previous step and separates the data into three 
CSV-s: one for singletons (entities with just one word), one for the first words
(all but the last word of multiword entities) and one for the last words (last
word of multiword entities)

import pandas as pd
ner_counts = pd.read_csv('ner_counts_5000_lemmas.csv')

ner_counts[:10]

Separate the entites to three

singletons_map = {}
last_word_map = {}
first_words_map = {}

import ast 

for row in ner_counts.itertuples():
    entity = ast.literal_eval(row.entity)
    if len(entity) == 1:
        curr_map = singletons_map.get(entity,{})
        curr_count = curr_map.get(row.label,0) 
        curr_map[row.label] = row.entity_count + curr_count
        singletons_map[entity] = curr_map
    else:
        last_word = entity[-1]
        curr_map = last_word_map.get(last_word,{})
        curr_count = curr_map.get(row.label,0) 
        curr_map[row.label] = curr_count + row.entity_count
        last_word_map[last_word] = curr_map
        
        first_words = entity[:-1]
        curr_map = first_words_map.get(first_words,{})
        curr_count = curr_map.get(row.label,0) 
        curr_map[row.label] = row.entity_count + curr_count
        first_words_map[first_words] = curr_map
        


singletons_map

first_words_map

# making first words mapping keys strings to allow convenient dataframe creation
first_words_map = {str(k):v for k,v in first_words_map.items()}

last_word_map

creating three dataframes, filling missing values with 0

single_csv = pd.DataFrame.from_dict(singletons_map,orient='index')
single_csv = single_csv.fillna(0)
single_csv

last_df = pd.DataFrame.from_dict(last_word_map,orient='index')
last_df = last_df.fillna(0)
last_df

first_df = pd.DataFrame.from_dict(first_words_map,orient='index')
first_df

Saving results to CSV files

last_df.to_csv('data/last_counts_lemma_5000.csv')

single_csv.to_csv('data/single_counts_lemma_5000.csv')

first_df.to_csv('data/first_counts_lemma_5000.csv')
