==> statistical_ner_labelling/notebooks/pipeline/3_3_sentences_from_given_last_word_list.ipynb <==
from estnltk import Text

last_word_list = ['jõgede','linnade']

from estnltk.taggers.system.rule_taggers import Ruleset, StaticExtractionRule

#create a ruleset from list
def list_to_ruleset(rule_list, len_check=None):
    ruleset = Ruleset()
    for word in rule_list:
        rule = StaticExtractionRule(pattern=word)
        ruleset.add_rules([rule])
    return ruleset

ruleset = list_to_ruleset(last_word_list)

text = Text('''On üsna kindel, et Eufrati ja Tigrise jõe ümbruses eksisteerisid linnad 2500 kuni 3000 eKr. Urukis elas sel ajal 50 000 inimest. Esimeste linnade rajamise aeg ja elanike arv pole selge.

Pole selge, kuhu paigutub linnade rajamine teiste innovatsioonide (loomade kodustamine, taimede kasvatamine ja kaubandus) kõrval. Traditsioonilise ettekujutuse järgi oli linnade tekke eelduseks toidu ülejääk põlluharimisest linna lähedal. Et aga osutub, et linnade ja kaugkaubanduse algus võivad olla palju varasemad, kui seni arvatud, on mõned uurijad selles eelduses kahtlema hakanud. On ka teooria, mille kohaselt linnad on tekkinud kaubanduse tõttu ja põlluharimine linnade toiduvajaduse tõttu. Mõned uurijad kahtlevad esimeste linnade tekke majanduslikes seletustes üldse ning peavad esimesi linnu eelkõige kuningliku või vaimuliku võimu keskusteks.

Tundub olevat selge, et esimesed linnad tekkisid jõgede lähedal, kus sai põlde niisutada. Niisutus vajas organiseeritust ja koostööd, mis nõudis tööjaotust. Esimesed suuremad linnad tekkisid tõenäoliselt Mesopotaamias ja Niiluse orus. Ka mujal tekkisid esimesed linnad jõgede (Indus, Huang He) ääres.

Pole selge, kas linnaelu levis esimestest linnadest või tekkis eri kohtades sõltumatult. Vähemalt Ameerikas tekkisid linnad nähtavasti Vana Maailma linnadest sõltumatult ja arenesid sarnaselt.''')

from estnltk.taggers.system.rule_taggers import SubstringTagger

subtagger = SubstringTagger(ruleset=ruleset,ignore_case=True)
subtagger.tag(text)

from estnltk_core.layer_operations import split_by_sentences

text.tag_layer()

sents = split_by_sentences(text,layers_to_keep=['terms'])

filtered_sents = []

for sent in sents:
    if len(sent.terms)>0:
        filtered_sents.append(sent)

filtered_sents

import pickle
with open('last_word_dataset.pkl', 'wb') as f:
    pickle.dump(filtered_sents,f)
